move several general functions to base class
